+++ omemo/__init__.py
@@ -1,5 +1,4 @@
 from .version import __version__ as __version__
-from .project import project as project
 
 from .backend import (
     Backend as Backend,