+++ CHANGELOG.md
@@ -13,6 +13,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Removed type annotations from enum members in accordance with a new mypy rule
 - Simplified the `JSONObject` type now that recursive aliases are properly supported
 
+### Removed
+- Removed project.py and simplified version.py as part of the migration towards pyproject.toml
+
 ## [1.2.0] - 15th of October, 2024
 
 ### Changed